Added flag, {{{SUMMARIZE_KILLS_RS_ENTRIES}}}.
In RROF, summarization is what scans objects in the remembered sets
(as opposed to the default generational collector, where cheney is
what scans such objects).

We can encounter objects in the remembered sets which actually have no
region-crossing pointers.  It would be good to remove such objects
from the remembered sets as soon as possible -- that is, while we're
scanning them while enumerating over the remembered set contents.
This flag enables that during the summarization scan.  (It also
filters out references to the static area, and anywhere else marked as
non-moving, which needs to be more cleanly separated.)

The flag is turned off by default, so this changeset on its own should
not introduce any change in behavior.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5921 c9681fa6-b8ae-40c0-a6da-707a35ad5973