Change to logic for whether we should force the refinement of the remsets.  Whether or not SMIRCY and RROF cycles are synchronized is not the (most) relevant issue for this decision -- what really matters is whether we're actually doing marking incrementally.
If the marking happens all at once (ie if the marking budget is
infinite, ie {{{BASE_BUDGET}}} is {{{-1}}}) then we will complete a
snapshot every time {{{smircy_step}}} is invoked, which means we do
''not'' want to refine the remembered sets every time we possibly
could, but rather delay for some number of collections (otherwise
we'll do a whole heap snapshot+refine on every collection, leading to
quadratic slowdown).  If the marking budget is finite, then it is
reasonable (though not a strict requirement) to refine every time the
snapshot is complete.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5922 c9681fa6-b8ae-40c0-a6da-707a35ad5973