Fix to performance problems introduced to GCBench yesterday.
changeset:5922 was supposed to be a pure refactoring with no
observable changes to the code base.  However, I did not understand
the semantics of the C preprocessor and thought that it could evaluate
boolean expressions like {{{(BASE_BUDGET == -1)}}}; it cannot, and so
we were reaching code that was not supposed to be evaluated and it was
causing severe performance issues.

This fixes the problem by replacing an attempted use of the C
preprocessor with an actual C expression instead.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5925 c9681fa6-b8ae-40c0-a6da-707a35ad5973