Second "fix" from yesterday and today -- this time to minor/major gc selection policy.  I observed that the former major/minor selection logic was bogus: it would often choose to perform major collections even when the to region was a ripe target for minor collection, because if the policy code said that a major collection is possible, it would ignore whether a minor collection is possible and attempt a major collection.  This led to an absurd number of unnecessary major collections.
The ''obvious fix'' of greedily performing minor collections instead
(by checking {{{can_do_minor}}} before {{{can_do_major}}}) does not
fix things immediately, however, because then the collector fills up
all of the regions with the results of minor collections and does not
have room to store the results of major collections.  (This is a sign
that the policy logic that states whether we can do major or minor is
broken -- possibly because the reserve region is not being accounted
properly in the policy code).

The quick fix I am adopting here attempts to strike a middle ground:
if the ''TO'' region can handle a minor collection and the successor
of ''TO'' can as well, then do a minor collection.  (This guards
against filling up the heap too fast.)  Likewise, if the successor of
''TO'' can handle a minor collection, then loop (to explore that
option) rather than doing a major collection into ''TO''.

This new logic is still not perfect, but it is a better approximation
of appropriate behavior, and it performs a bit better on the
benchmarks.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5937 c9681fa6-b8ae-40c0-a6da-707a35ad5973