Third "fix" from my investigations on Jan 18th and 19th -- this one directly affects the heap expansion policy (which indirectly affects the sequence of choices between minor/major collections).
This changeset follows through on a suggestion that I made in the log
entry for changeset:5469:

  "It would probably be better to precisely calculate the number of
   promotions out of the nursery and then only subtracting that from
   the live storage in the region. "

----

Back in changeset:5468 and changeset:5469 I decided to tinker with the
estimate of total heap size that guides the heap expansion policy.

Each region carried how many bytes were the result of the last major
collection into that region.  Originally I thought that summing these
would provide a reasonable estimate of the total heap size when we
were in a steady state of computation.  But this ended up being an
over-estimate and led to over-aggressive expansion of the heap -- I
documented one hypothesis for why this happens with changeset:5468,
where I attempted to correct or the over-estimate by subtracting the
nursery size.

But that correction is too aggressive; changeset:5469 was a band-aid
to avoid some truly absurd results.  This changeset is a better
band-aid.  I probably should be thinking harder about whether this
method of estimating the total heap size is actually sane -- at the
very least, I should be more serious about measuring how much error
this estimation introduces.

----

With this changeset (the third of three conceptual changes I found
over the weekend), the {{{5twobit:long}}} benchmark has a saner ratio
of minor/major collections; there are almost no runs of minor-gc's
with length less than 2.  (When I started on this investigation, there
were 246 such runs for {{{5twobit:long}}}.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5938 c9681fa6-b8ae-40c0-a6da-707a35ad5973