Added procedures to measure how large the smircy stack has grown.
{{{smircy_stack_size}}} measures how many entries the stack has grown
to, ''including'' zeroed-out entries that have not yet been popped off
the stack.

{{{smircy_stack_count}}} measures the number of non-zero entries on
the stack.  As long as smircy uses a {{{MARK_ON_PUSH}}} policy, this
will correspond to the size of the set of entries that remain to be
processed.

When {{{smircy_stack_size}}} grows far beyond {{{smircy_stack_count}}}
that is a sign that something is going terribly wrong.  (I saw some
truly awful ratios earlier today; hopefully the fixes I have found
will bring things down to an acceptable level.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5940 c9681fa6-b8ae-40c0-a6da-707a35ad5973