When there's only one region, there's (almost) no reason to initiate a mark/refinement (because after each gc, there should not be any elements in the region's remembered set; nothing to refine).
I say "almost" because the logic above discounts the presence of the
static area, which has its own remembered set which may indeed have
elements that could be refined.  But that is not an event we expect,
and I am pretty certain it is not worth the cost of starting up a
mark/refinement.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5947 c9681fa6-b8ae-40c0-a6da-707a35ad5973