Changing code so that each mark/refinement increment has its work bounded by a function of the words promoted out of the nursery during the previous collection.
As part of this, I am turning {{{DONT_USE_REFINEMENT_COUNTDOWN}}} on;
so now there is no "forced finish" after N collections.  The
mark/refinement is given its increments to run, and presumably it will
finish in a reasonable amount of time.

Technically, the work bound for mark/refinement can be a function of
three parameters: {{{P}}}, {{{C}}}, and {{{F}}}, where:

  {{{P}}} : is the number of words promoted out of the nursery

  {{{C}}} : is the number of words allocated in the nursery but not promoted

  {{{F}}} : is the mark refinement parameter 
  (the field {{{rrof_refinement_factor}}}, set by 
  the {{{-refinement}}} command line switch)

The current bound calculation ignores {{{C}}} entirely, and uses
{{{P}}}/{{{F}}} as the bound on mark/refinement work.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5949 c9681fa6-b8ae-40c0-a6da-707a35ad5973