More fields to track promotion information.  I am not sure yet if these are strictly necessary, but I do not yet know how else I am going to gauge the rate of promotion.
The constaints Will and I have been discussing imply that we want a
collection cycle to complete before too much promotion has occurred
since the last completed snapshot, but the time that a collection
cycle will complete depends on how often one does major collections --
that is, the selection between major and minor collections seems like
it should depend on the last snapshot size, how much was promoted
during the snapshot development, and the promotion rate for the
remainder of the cycle.

I need to talk to Will about this some more.  But for now I can at
least gather data on what the average number of words per promotion.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5966 c9681fa6-b8ae-40c0-a6da-707a35ad5973