Simplifying policy for shifting the tospace; the previous code is a hold-over from when the the collection polciy was less sane.
In the current setup, the regions after the the to-space region are
all supposed to be empty (at least up to the next region undergoing
major collection; that one is obviously usually non-empty).  If you
ever encounter a non-empty region after the to-space region, it must
have been skipped for being popular, and should be skipped over during
the to-space selection.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5976 c9681fa6-b8ae-40c0-a6da-707a35ad5973