GCC actually seems to ``optimize'' the check {{{ assert( (s->min_real > 0) && (s->min_real+1 < 0) ); }}} to {{{ assert( FALSE ); }}}
See http://www.fefe.de/intof.html for some independent confirmation of
this matter.  That page also points out that the technique here for
calculating the maximum value is flawed: shifting a negative value is
undefined in C.  But that's not a problem I am going to try to address
right now.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5997 c9681fa6-b8ae-40c0-a6da-707a35ad5973