To reduce fragmentation (at least as reported by the Waste entry of {{{memstats}}}), increasing pool size to 1024.
Experiments on my laptop indicate that 1024 is a sweet spot that tends
to reduce both the reported Rts size and the reported Waste.  Going
above 1024 can further reduce the reported Waste, but the Rts size
increases, which indicates to me that at that point one is trading
external fragmentation (introduced by our allocator's working to align
blocks) for internal fragmentation where many of the pools within
{{{summ_matrix.c}}} have unfilled entries.

I am not certain why the {{{gclib_alloc_rts}}} uses the aligning
allocator in the first place; my hypothesis is that any entries in the
{{{gclib_desc_b}}} (and {{{gclib_desc_g}}}) tables must be aligned
addresses.  (This then raises the question of whether I need to be
using memory with state stored in such tables... )


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6065 c9681fa6-b8ae-40c0-a6da-707a35ad5973