"Eliminate" fragmentation introduced by mark bitmap by being willing to use the extra space introduced by rounding up to the page boundary.
In my experiments with 5nboyer:5, this made no difference; so perhaps
this will not help memory or time usage.  But at least with this
change the extra space allocated will be charged to the Rts (and to
Smircy once that is accounted for) rather than being grouped in with
all other Waste due to {{{alloc.c}}} fragmentation).


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6068 c9681fa6-b8ae-40c0-a6da-707a35ad5973