Code review revealed fragility that I'd rather just assert-fail on.
The fragility: If we're looking at an address that is outside the
bounds produced by {{{allocate_bitmap}}}'s invocation of
{{{gclib_memory_range}}}, then something is deeply wrong (in a way the
state maintained by the {{{smircy_context_t}}} is not going to be
recover from).

So instead of doing {{{min}}}/{{{max}}} guesswork, check that the value
produced by {{{gclib_memory_range}}} is sane, and then use it.

(One could conclude from this that {{{expand_bitmap_core}}}'s
interface could be simplified by removing its second and third
parameters; but at one point I was trying to avoid expanding to cover
all of {{{gclib_memory_range}}} and maybe at some point in the future
I'll return to such a strategy; so the parameters will stay as is for
now.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6069 c9681fa6-b8ae-40c0-a6da-707a35ad5973