Will's experiments exposed an over-sight in how I was budgeting summarization in the face of new regions being added on the fly.
----

Here's an invocation one can run to expose the problem:
{{{
% ./larceny -rrof -size1 4M -load_hard 5 -popularity 10 -sumzcoverage 3 -sumzbudget 2.2 
> (load "queue2.sch")
> (queue-benchmark 1000 1000000 25)

--------------------------------------------------------
queue1000:1000000:25
Larceny Panic: Sys/summ_matrix.c;3153: Assertion failed.
}}}

----

In essence, the relevant questions here are all of the flavor: where
should a newly allocated region go with respect to the incremental
summarization.  Should it be considered summarized?  Should its
remembered set be considered scanned?  (These are orthongonal issues.)

Guiding the answers to these questions is the constraint that pause
times must remain bounded by a constant; therefore, increasing the max
number of remembered sets that we are allowed to scan during a
summarization step would be breaking the rules.

I believe I have came up with a reasonable resolution (still to be
implemented), but that's not what this changeset is implementing; all
this changeset is an eager assertion failure when we discover that
we've put ourselves into a state where the budgeted state might fail
to support the next wave of summarization.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6073 c9681fa6-b8ae-40c0-a6da-707a35ad5973