Fix for problem described in log entry for changeset:6073.  (This is the "resolution" I alluded to the log entry.)
The simple story is: newly allocated regions are treated as
unsummarized, but ''already'' ''scanned'' for their contributions to
the summaries under construction.  This is sound, because newly
allocated regions are empty.

I considered implementing this treatment of the new regions by
assigning them a {{{gno}}} less than all the others, and thus the
{{{rs_cursor}}} would be past them.  However, that would have been a
ridiculous pain to implement.  It was ''far'' easier to continue
assigning the new regions a {{{gno}}} greater than all the others
(besides the static area's {{{gno}}}) and just change the remset
scanning code to go from the end remset and works its way to the
front, which is what this changeset does.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6074 c9681fa6-b8ae-40c0-a6da-707a35ad5973