Got around to investigating failure discussed in log for changeset:6072.
The failure was exposed by interactions between changeset:6068 and
changeset:6069.  Attempting changes like page-aligning the size of the
bitmap requires a little bit more work than what I did in
changeset:6068.

In this case, the address range covered by the heap (and acquired via
{{{gclib_memory_range}}}) can shrink as well as grow (that is, it is
not monotonically growing).  It is simplest for me to force the mark
bitmap's address range to monotonically grow, and some parts of the
code reflect this simplification.  But I was not actually doing it
consistently in my code; {{{allocate_bitmap}}} did not consume
parameters representing the old address range.

This changeset makes the code saner (but still does not enable
page-alignment of the mark bitmap).  Now {{{allocate_new_bitmap}}}
consumes an "old" range as well as producing the new range.
{{{allocate_bitmap_expanding_from}}} is an alternative to
{{{allocate_bitmap}}} for invoking {{{allocate_new_bitmap}}}.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6075 c9681fa6-b8ae-40c0-a6da-707a35ad5973