Added code to support better handling of requests for 8K blocks (to reduce fragmentation).
Switch {{{PARCEL_8K_BLOCKS}}} to {{{0}}} to disable the change.

This is an approximation of a strategy suggested by Lars in his
original comments; instead of allocating individual 8K blocks (which
can introduce up to 4K of fragmentation -- boo), allocate a large
16*8K (= 128K block) and parcel out individual 8K blocks from that.
(16 is a parameter in the code; see {{{NUM_8K_BLOCKS_PER_ENTRY}}}; it
might be worthwhite to increase it to something closer to the word
size).

(This change can actually introduce *more* fragmentation than we had
before, since a single 8K block can cause 120K of extra space to be
held.  But that extra space should be reused to satisfy future
requests, so in pratice this change will be a win.)

Anyway, this change reduces the Waste on RROF 5earley:13 from
6,827,904 Megawords down to 1,492,992 Megawords; I also see evidence
that this helps the elapsed times (not the CPU time), which indicates
to me that the change is reducing the number of page faults.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6082 c9681fa6-b8ae-40c0-a6da-707a35ad5973