This is much saner than what I was doing before.
The earlier code was attempting to emulate the standard math trick of
making some representation for postive infinity the initial value for
an accumulator (that will be using {{{min}}} as its combining
operator).  While that is a nice general approach, attempting
to implement it in C leads to craziness like this:
  {{{(int)(((unsigned)~0) >> 1)}}}
'''and''' the approach is both unnecessary and undesirable here.

We ''have'' the window size already available, and ''that'' is 
the appropriate initial value for the accumulator; because no 
operation can conceptually extend beyond the window anyway!

(Well, another semi-reasonable initial value for the accumulator is
the window's {{{size + 1}}}.  But I think this will be saner.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6087 c9681fa6-b8ae-40c0-a6da-707a35ad5973