Added accounting of memory usage by summary sets and snapshot marker.
A big changeset (for me at least), but most of it is semi-standard
boilerplate for our awkward statistics gathering-and-transmission
system.  That is, the additions for summary sets (aka ``summ'' or
``summsets'') and snapshot marker (aka smircy) are roughly analogous
to code that was already present for tracking memory used by the
remembered sets.

Something important that this changeset does '''not''' fix: some of
the allocation that should be charged to the summary sets is (still)
being charged to the remembered set.  I will be fixing that in the
near future, but I had to actually get some report of the summary sets
memory usage in place before trying to reassign to whom the relevant
hash tables are charged.

----

Also, I put in some state to break down the memory usage at the peak
of the computation.  Reporting max words for each part alone (as we
have been doing and continue to do even with this changeset) leads to
a situation where the whole can be less than the sum of "its parts."

The max words for each part of memory is certainly interesting
information to see, but confusing/difficult to relate to the max mem
words in general.  So at some point I'll probably finish off that
change and make {{{run-with-stats}}} report the break-down at the peak
of total mem words in addition to (or ''instead of'') the max values
of each part.

----

FYI: there is definitely some problem when one uses v0.97b1 to host a
build of the {{{incrsumz/}}} branch (I just did it on artichoke on a
clean checkout of {{{incrsumz/}}}, which is what has delayed my
attempt to check this in.  I am not sure how the poblano build got
corrupted in a similar manner; I may have done a build on Feb 13th
using a private development build rather than the system installed
Larceny.  Or maybe the fact that the problem is exposed via v0.97b1
rather than v0.963 is a red herring.

But in any case, while in {{{branches/incrsumz/larceny_src}}} on
{{{artichoke.ccs.neu.edu}}}, with a clean checkout:

{{{
% larceny -- setup.sch -e "(setup-load-build 'sassy)"
Larceny v0.97b1 (beta test) (Aug 25 2008 12:37:27, precise:Linux:unified)
Loading Larceny compatibility package.
src/Compat/Larceny/compat2.sch
  ...
Compiling File (with Syntax) lib/SRFI/srfi-69.fasl
larceny.heap, built on Mon Aug 25 12:39:23 EDT 2008
> (exit)
% ./larceny.bin -stopcopy -- src/Build/iasn-larceny-heap.fasl
  ...
Welcome. Type (help) for help.
Install twobit's macro expander as the interpreter's ditto
Replace and populate the top-level environment
src/Lib/Common/toplevel.fasl
src/Lib/Arch/IAssassin/toplevel-target.fasl
done with load
WARNING from macro expander:
Redefining 
case-lambda


Error: fxlogand: 268435448 is not a fixnum.
Entering debugger; type "?" for help.
debug> b
=> compiled procedure debug/enter-debugger
   system continuation
   compiled procedure lgand
   compiled procedure #f
     ...
   compiled procedure #f
   compiled procedure sassy
     ...
debug> 
}}}

in other words, this is some sort of problem internal to the compiler;
from what I've been able to see, it seems to be some sort of weird
problem with the execution of the {{{fxlogand}}} primop...  other than
that, I do not know what's going on.  I cannot take the time to
investigate it more thoroughly at the moment, but one should be able
to do a checkout of the v0.97b1 code (on the {{{trunk/}}} from August
25th) and replicate the problem by attempting to build this code base
(on the {{{incrsumz/}}} branch)



git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6090 c9681fa6-b8ae-40c0-a6da-707a35ad5973