Fix for incremental marking bug exposed by GCBench that has had me stumped for days.
History: 

 * Will reported the bug on February 25th.
 * I determined it was likely to be the incremental marker 
 while working in Will's office over the last couple of weeks.
 * With that information, last weekend I did a search over 
 the {{{-refinement}}} parameter space and managed to 
 replicate the bug on my own machine using the following 
 parameters:
{{{
% larceny -rrof -refinement 5 -sumzbudget 2 -sumzcoverage 2 -popularity 8 -size1 4M -- gcbench3.fasl 
...
> (gc-benchmark 22 5)
}}}
 * Then yesterday and today I instrumented the incremental
 marking code and finally tracked down the actual bug.

The essential problem was that an object was being inserted into the
mark stack (ie the frontier for the incremental marking) and
subsequently lost.  The object in question is a large object, which
are handled specially by the mark stack, and a rare control path with
a mark stack management bug was being exercised.

The bug itself was: the mark stack is represented by a list of
segments (we allocate the segments on demand, but attempt to reuse
them during the course of marking), and when a los stack segment was
''reused'' during a push, it was not hooked up with the rest of the
los stack (linked list).

As a way to better exercise the code base, I should probably try
reducing the {{{GC_LARGE_OBJECT_LIMIT}}} in {{{src/Rts/Sys/memmgr.h}}}
to something small enough that these kinds of things are more likely
to get exposed; dividing the limit by 8 or 16 may be a simple change
that would probably increase the number of control paths exercised...


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6144 c9681fa6-b8ae-40c0-a6da-707a35ad5973