Further revisions of FFI chapter.
Cleaned up introduction.

Added spaces to certain line endings in attempt to improve output
formatting.

Revised FfiForeignValues section to use bulleted lists with a little
bit of nesting.  (One might hate bullets but it seemed like most
reasonable way to present the categories of questions.  If one is to
eliminate the bullets, I would say the right way to do it would be to
revise the choice of presentation entirely, which I do not want to
spend further time on).

Added grammar as my way to describe foreign type descriptor
"expressions."

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6224 c9681fa6-b8ae-40c0-a6da-707a35ad5973