Added {{{render-memory-usage}}} and started thinking about {{{plot-memory-usage}}}.
However, the very first invocation I made of {{{render-memory-usage}}}
brought up the problem that the sum of each components max memory
usage can exceed the max total memory usage.  So I am going to switch
gears now and try to add max-at-peak stats for each component (since
it is much easier for me to add new entries to the stats s-exp now
that I know about the {{{stats-dump-on}}} and {{{stats-dump-stdout}}}
procedures.

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6270 c9681fa6-b8ae-40c0-a6da-707a35ad5973