Revised code so that:
 - the width of a bar in a singleton group is not infintessimally thin

 - one width is used for all of the bars in a plot (which helps avoid
   mis-interpretation of the data due to the human eye interpreting
   area as relevant when it is the actually the height that is the
   important thing.

Also, noticed that my shift to tracking memory usage at peak rather
than max overall has led to a very important conclusion: the memory
lossage due to fragmentation at *peak* occupancy is often not
significant even if the max fragmentation was itself high.  In
particular, on nboyer it is clear that to reduce memory usage, I
should focus on the remembered set and the marking bitmap, *not* on
the fragmentation and it looks like not on the summaries either (at
least for now).

It is still important to know that the max overall for the other
components is high, since the above conclusion does not imply that
fixing those two components will actually make the max memory usage go
down by the corresponding amount.  I may think about ways to try to
incorporate both pieces of data into my rendering...

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6276 c9681fa6-b8ae-40c0-a6da-707a35ad5973