Decided the output of {{{render-mmu2}}} is interesting enough to warrant reuse; gave it its own data definition.
Factored {{{plot-mmu2}}} into smaller pieces; in particular, the
{{{plot-mmu2/core}}} procedure is useful if you already have a
rendered mmu in hand.

Added functions to turn the benchmark descriptions into their rendered
mmus.  The unstarred variant folds all of the data together into one
rendered mmu; the starred variant keeps them separated in a list.

All of these compose to plot the observed MMU.

For example, to see it for the benchmarks individually:
{{{
(for-each plot-mmu2/core
          (bench-descriptions->rendered-mmu2*
           (describe-stats-of-logfile ".*May29.*")))
}}}

To see the MMU for all combined:
{{{
(plot-mmu2/core
 (bench-descriptions->rendered-mmu2
  (describe-stats-of-logfile ".*May29.*")))
}}}

(The above example are expressions I just ran today; they make use of
the regexp matching that {{{describe-stats-of-logfile}}} employs if it
cannot find a file by the given name.)

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6281 c9681fa6-b8ae-40c0-a6da-707a35ad5973