Revised representation of full cycle cursor (earlier rep was simulatenously counting up and down without remembered how large the original total was, which leads to trouble); renamed fields involved to clarify their role.  Also revised how the total major-collections in full cycle is initially computed; old code over-counted because it included empty regions (the new calculation is factored as {{{nonempty_region_count}}} function).
When determining if an attempt to do a summarization step would no-op,
use the new region count calculation.  (I use a dynamic computation
rather than the number of regions at the start of the summarization
cycle because allocation taking place during the summarization cycle
changes the number of regions that need to be summarized; another
option could be to use the upper bound on the maximum number of
regions that we could possibly need to scan during the summarization.
I might switch to that strategy in the future.)

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6292 c9681fa6-b8ae-40c0-a6da-707a35ad5973