Extended summ matrix interface to pass along dA (it will be used as corrective factor in calculation of remset scan rate to account for max concurrent allocation).
To support calculation of the parameter above, I need to know the
allocation target and the region size; there was already a field in
memmgr_internal.h (which defines fields of {{{DATA(gc)}}}) for the
allocation target, but the region size was not being tracked (this may
seem surprising, but its a consequence of my attempt to continue down
the path of allowing one to set the sizes of the regions via command
line options).  Since there may be a range of sizes, I will track the
min and max, and then use the appropriate one based on the needs of
the context of use.

I also had to add code to actually '''maintain''' the allocation
target (as a field in the {{{DATA(gc)}}}).  I added this code to
{{{rrof_completed_regional_cycle}}}, but doing that in turn required
that the potentially modified heap sizes get synchronized before the
{{{rrof_completed_regional_cycle}}} invocation.

Finally, computing the allocation target requires using much of the
same code that is currently used within {{{rrof_gc_policy}}}, so I
factored the common portions out into a separate procedure,
{{{rrof_calc_target_allocation}}}.

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6294 c9681fa6-b8ae-40c0-a6da-707a35ad5973