Fixed bug in {{{nasm-disassemble-bytevector}}}.
At some point, {{{call-with-output-file}}} changed to open a textual
output port, and the invocations of {{{write-char}}} within
{{{nasm-disassemble-bytevector}}} were not writing raw bytes but
instead prepending junk that messed up the correspondence between the
bytevector's contents and the contents of the output file.

This change fixes that by using the new procedures defined in the R6RS
for the binary output.

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6303 c9681fa6-b8ae-40c0-a6da-707a35ad5973