Added new method for enumerating the ''locations'' within the remembered set (rather than the ''objects'').
With the scanning function looking at individual locations within
objects, it is not immediately clear how to code up the action of
refining the remembered set during the traversal (at least not without
passing along more context to the scanning function, and complicating
its interface to some degree).  Since the regional collector will be
traversing immediately disposable summaries rather than remembered
sets, this is not really relevant and so I left it out of the
interface.

The current implementation itself is mostly trivial, the bulk of it
dispatching to the pre-existing object-enumerators with a wrapper,
{{{apply_f_to_summary_loc_entry}}}, that invokes the location scanner
on each location within object passed along by the object enumeration.

A future changeset should revise the {{{summary_t}}} interface to be
able to enumerate locations rather than objects, and then the code of
{{{enumerate_remembered_locations}}} will need to be revised
accordingly.

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6318 c9681fa6-b8ae-40c0-a6da-707a35ad5973