Free {{{vector}}} procedure's field initializions from write barrier.
With Will's prompting, Felix convinced himself that replacing the
{{{.vector-set!:trusted}}} with {{{.vector-set!:trusted:nwb}}} is
indeed sound in this context.

It is valid because:

 1. When control returns from the {{{(make-vector ?n ?t1)}}} expression,
    the resulting vector is in the nursery (any gc that takes place to
    make room for it happens before the vector is actually allocated).

 2. The series of assignments into the vector via
    {{{.vector-set!:trusted}}} are not going to cause a
    collection:

    a. the {{{?t2 ...}}} expressions are all (simple) variable reference
       expressions

    b. the {{{?n2 ...}}} expressions are all uniform arithmetic
       calculations that will be reduced at compile-time to constant
       references.  (This was a sticking point for Felix.)

Felix still wanted to sanity-check this, so his first draft of the
replacement also included the lines marked with {{{XXX}}}, which
double-checked that a collection had not occurred in the range of
interest.  However, including that assertion slows things down
('''way''' more than the speedup from eliding the write-barrier), so
those changes are commented out in this changeset and should be
removed entirely in a future changeset.

----

Finally, some rough results regarding this:

Felix did runs of queue1000:1000000:10 ''(note *not* :50)'' on the
trunk, and on RROF with and without this change.  Felix found that for
his single run of that benchmark, eliminating the write barrier does
improve mutator time, but not enough to leap-frog the entire
performance difference compared to the trunk.

Then again, the performance differential from the trunk that Felix
sees (for that benchmark on Felix's machine) does not seem as
significant as Will is seeing (with a different benchmark and
different machine) so this is really not terribly informative.

  EXECUTIVE SUMMARY:
  ------------------
  2GEN TRUNK                           ==> 29.26 mutator time (sec)
  RROF WITH SIMPLEST VECTOR            ==> 35.96 mutator time (sec)
  RROF WITH VECTOR:NWB (non asserting) ==> 33.42 mutator time (sec)

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6355 c9681fa6-b8ae-40c0-a6da-707a35ad5973