More robust fix for Ticket #617.
I reverted changeset:6132 because that introduced a bigger problem
than it solved.  (In particular invocations of {{{build-runtime}}}
were not rebuilding the runtime, for precisely the reason documented
in comments added with changeset:5803 -- we ''want'' to conservatively
invoke {{{make}}} unconditionally from {{{build-runtime}}} so that we
do not have to put the runtime dependency information into
{{{makefile.sch}}}).

So, rather than avoid the "double-make invoke" problem documented in
Ticket #617, I am changing the generated Makefile to be more robust in
the face of double-make invocations.

The trick is this: I added an artificial dependency, {{{kill_version_o}}},
to the {{{larceny.bin}}} entry in the {{{Makefile}}}, and placed the
target before all other dependencies.  The principle here is: If we
initiate a runtime build, then we will first commit ourselves to
'''do''' the build of {{{larceny.bin}}}, regardless of what timestamp
the initial {{{version.o}}} has.

With this change, the removal of {{{version.o}}} ''after'' the runtime
build might be unnecessary.  But it seems to me that it is safest to
leave that be.

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6357 c9681fa6-b8ae-40c0-a6da-707a35ad5973