Large ugly hacked-on band-aid.
When enumerating over the remembered set, the scanner might allocate
new regions, which will change the gno of the static area.

I had thought that the idea in changeset:6405 [1] would side-step the
problem by handling the static area before looking at the rest of the
remembered set (via a small revision of the control flow), but all
that did was make the problem come up less frequently!

In the end, even with the you still have two coroutines: the
enumerator, which is often attempting to pass along what it thinks of
as an unchanging {{{gno}}} for the static area, and the cheney
scanner/forwarder, which will occasionally allocate a region (which
then modifies the {{{gno}}} of the static area).  The point is that
the enumerator is (currently) wrong to think that the desired {{{gno}}}
is a constant.

The hack applied here is to treat the static area's {{{gno}}} as a
special case and pro-actively re-extract it within the remset
enumerator traversal.

(This is not what I want to do in the long term, though.  At this
point, I think a better overall solution is to redo the {{{gno}}}
assignment policy so that the static area's {{{gno}}} is constant,
just like the {{{gno}}} of all the other regions.  This can be done
e.g. by giving the static area a {{{gno}}} of {{{-1}}}, or {{{1}}},
or anything that is set aside as ``special.'')

[1] Note that changeset:6405 was itself quite buggy; changeset:6408
    and changeset:6413 contain some revisions that made it more
    robust, but they too fall victim to the fundamental mistake I made
    here.

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6414 c9681fa6-b8ae-40c0-a6da-707a35ad5973