Added code to the SSB overflow handler to schedule a garbage collection if mutator activity exceeds its bound.
This approach is not quite right in a couple of ways:

  -- It only schedules a collection (by artificially filling the
     nursery); it does not force one to immediately occur.  There may
     exist worst-case mutators that perform assignments in a loop
     without allocating stack or cons'ing; this approach will not
     catch them (and thus they could cause invalidate all of the
     summary sets before any collection occurs).

  -- It probably should not wait for mutator activity to exceed
     the bound before acting.

Also, there are issues like the fact that {{{calc_cN}}}'s current
implementation has obvious potential performance bugs, and therefore
it may be a bad idea to invoke it from the SSB overflow handler.

But this is a start.

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6427 c9681fa6-b8ae-40c0-a6da-707a35ad5973