Placate more GCC warnings by not falling off end of void function.
Would be better to fix interface of {{{uremset_t}}} to either assign a
saner semantics to the return value, remove it entirely, or provide a
different (ie optionally employed) interface for extracting this
information when available.  But I do not want to take the time for
any of that right now.

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6459 c9681fa6-b8ae-40c0-a6da-707a35ad5973