GCC warnings suggested added parentheses around + or - inside shift.
I am horrified that I left this expression like this with a
straight-face in the first place.

I believe the logic is supposed to be just like how the first partial
word is set up a few lines above, except of course a direct analogy is
broken because you want the exact opposite (in that you want to shift
up a bunch of zeros to be OR'ed with the high bits, instead of
shifting up a bunch of ones.)

Perhaps the most digusting thing is that + 1 mixed in there.  "Didn't
you just decrement limp, up above?"

If this code was actually something that was part of critical
infrastructure, I'd be tempted to tear this out and put something in
that at least documented the intermediate states of the computation,
and preferably was a self-documenting expression.  (As it is I am
tempted to add an assert fail to the code to make sure no one uses
this without first double-checking the logic...)

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6463 c9681fa6-b8ae-40c0-a6da-707a35ad5973